+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+- fix: update replica message status getter for new replica mapping
+
 ### 2.0.0-rc.17
 
 - chore: bump configuration to v0.1.0-rc.23